+++ src/module/data/item/ability.mjs
@@ -562,6 +562,8 @@ export default class AbilityModel extends BaseItemModel {
 
         messageData.system.parts.push(rollPart);
       }
+    } else {
+      DrawSteelChatMessage.applyRollMode(messageData, "roll");
     }
     // TODO: Figure out how to better handle invocations when this.actor is null
     if (resourceSpend) await this.actor?.system.updateResource(resourceSpend * -1);